// src/FE/src/api/generated/models/customerListItemViewModel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface CustomerListItemViewModel {
  id: string;
  /** @nullable */
  customerNumber: string | null;
  name: string;
  /** @nullable */
  address: string | null;
  /** @nullable */
  zipCode: string | null;
  /** @nullable */
  city: string | null;
  /** @nullable */
  country: string | null;
  /** @nullable */
  email: string | null;
  /** @nullable */
  contactPerson: string | null;
  /** @nullable */
  phone: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  jobCount: number | string;
  isFavorite: boolean;
}