// src/FE/src/api/generated/models/customerDetailViewModel.ts
/**
 * Generated by orval v8.16.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { CustomerJobViewModel } from './customerJobViewModel';

export interface CustomerDetailViewModel {
  id: string;
  name: string;
  address?: string | null;
  email?: string | null;
  contactPerson?: string | null;
  phone?: string | null;
  jobCount: number;
  jobs: ReadonlyArray<CustomerJobViewModel>;
}